requirements.txt
aniso8601==9.0.1
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.5.0
awscli==1.22.101
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
boto3==1.21.46
botocore==1.24.46
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dell-recovery==0.0.0
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
docutils==0.15.2
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.6.0
Flask==2.1.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
monotonic==1.5
netifaces==0.10.4
numpy==1.22.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.9
PAM==0.4.2
pandas==1.4.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.1
progressbar==2.5
protobuf==3.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyPDF2==1.26.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.7.2
s3transfer==0.5.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sos==4.2
SQLAlchemy==1.4.35
sqlparse==0.4.2
systemd-python==234
tk==0.1.0
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.14.0
wadllib==1.3.3
Werkzeug==2.1.1
xkit==0.0.0
zipp==3.8.0
